Create AddApplicantPageController and route, move AddCourseActionController to FormActions
