+++ app/routes.php
@@ -17,5 +17,6 @@ return function (App $app) {
     $app->get('/admin/courses/add', \Portal\Controllers\Pages\AddCoursePageController::class);
     $app->post('/admin/courses/add', \Portal\Controllers\FormActions\AddCourseActionController::class);
     $app->get('/admin/cohorts', \Portal\Controllers\Pages\CohortsPageController::class);
-    $app->post('/admin/applicants/add', \Portal\Controllers\Pages\AddApplicantController::class);
+    $app->get('/admin/applicant/add', \Portal\Controllers\Pages\AddApplicantPageController::class);
+    $app->post('/admin/applicant/add', \Portal\Controllers\FormActions\AddApplicantActionController::class);
 };
