+++ templates/applicants.phtml
@@ -22,7 +22,7 @@
     <div class="container">
         <div class="applicant-header-container">
             <h1>Applicants</h1>
-            <a href="applicants/add">Add applicant</a>
+            <a href="applicant/add">Add applicant</a>
         </div>
 
         <?php if (count($applicants) === 0): ?>
